Archer Base Attack (#52)
Plugs in a new kind of Action: LaunchProjectile. LaunchProjectile creates a networked GameObject that then can do its own logic, with its own scripts. Which projectile is created is determined by a "SpawnedObjects" field in the Action's ScriptableObject. In our case we've made an "Arrow" prefab with the following scripts:

ServerProjectileLogic
NetworkProjectileState
ClientProjectileVisualization.

I have also generalized ClientCharacterMovement to ClientGenericMovement, and shared the "smoothing" code between ClientCharacterVisualization and ClientProjectileVisualization. I expect that this movement logic will be considerably simplified once MLAPI is handling movement as an internal feature.