Feature/Boss trample attack (#62)
This implements the boss imp's "charge attack". For now, this is the boss's basic-attack (not a specially-triggered attack, because we don't have special attacks yet).

To avoid changing the scene just to test this, I added a new temporary testing command: press "B" on the server/host to spawn a boss imp. (You can still press "E" to spawn a normal imp.)

Notes:

Because all character's rigidbodies are kinematic (and need to be, because they're moved manually to correspond with nav-mesh-determined positions), I needed to enable kinematic collision callbacks to know when the boss runs into someone. (To do this, I chose Edit->Project Settings->Physics, and set "Contact Pairs Mode" to "Enable Kinematic Kinematic Pairs".) This is why DynamicsManager.asset needed to be checked in.
There was a "missing scene" in the editor build settings; I replaced this missing scene with DungeonTest scene so that I could make builds. This is why EditorBuildSettings.asset was checked in.
Also fixed an unrelated error in Imp base attack -- the version in develop is bugged so that the imps can only attack once, ever. (Needed to fix for testing, forgot to revert.)
The boss's "I'm going to charge" animation is not yet ready, so right now the boss just performs a base-attack animation (swinging his hammer) prior to charging. I expect there will also be a custom movement animation (or maybe just particles) while the boss is sprinting, but right now he just runs.
Similarly, the ground-arrow is a programmer-art placeholder graphic.
I made the boss's Collider a bit bigger so he can trample better.
Edit: the boss's trample animations are available now, so I hooked them up. There's a new animator trigger "Trample" for this purpose