Mage base attack (#60)
Implement's the mage's base attack, which is a "fake projectile" attack (no physics; always hits). In order to test the mage, I changed the default character-class to be a mage. (There's a variable for this on the BossRoomState prefab; it's the only change to that prefab.)

To fully support this sort of attack, as well as projectile-attacks in general, I also changed the Action system in two major ways:

Actions can now be "non-blocking" for a part of their duration (the part after ExecTimeSeconds has elapsed). This lets mages perform other actions while their mage-bolts are in the air.
Actions are now sent in groupings called ActionSequences, rather than the parts of a sequence being sent one at a time. This allows the ActionPlayer to make more intelligent choices about whether a sequence should be played immediately, enqueued, or discarded. The purpose: so we can behave correctly when the player spam-clicks on enemies! Now, if the server receives a request to play a sequence that's identical to what it's already playing, it can queue it up.
Other notes:

As a side-effect of the new enqueuing ability, I needed to ensure that attacks actually end. This is why a bunch of unrelated attack actions were changed. (The only change was to their DurationSeconds, changing it from 0 to 1.) I think the previous value of 0 was basically a "bug that happened to work".
The "Chase" and "Revive" action data was only changed because I added a value to the middle of an enum, and it screwed up the values they were showing. (In other words, they effectively have the same values as before.)
All projectile artwork is placeholder programmer-art.