Action Targeting and Healing (#81)
This PR includes a new Target Action, as well as the Mage's "Skill2" Heal Effect, to take advantage of it.

A TargetAction runs persistently on the server until your target is cleared.

A TargetActionFX runs persistently on the client, and manages target reticule creation. New reticules are only instantiated in cases of ill fortune (the reticule's parent being destroyed).

ClientInputSender has been reworked to better accommodate skill1/skill2, as well as having different input triggers. An input triggered from a mouse click should cast a ray to get the target, but an input triggered from the UI or keyboard should rely on the character's pre-existing target, if any.

Made Movement in ClientInputSender also driven from FixedUpdate. This is because movement and targeting were conflicting. Movement requests are also only generated every 0.5 seconds when holding the button down, as an efficiency gain.

New Heal Action has been added that just uses the MeleeAction logic (hooray for logic reuse!) The MeleeAction has been slightly retooled to use a new "IsFriendly" bool on the ActionDescription. If true, PCs will search for PCs to hit, and NPCs will search for NPCs (of course if the action has a set target, that is preferred if at all possible).

Fixed a bug with ActionPlayer.ClearActions (was not actually clearing the non-blocking Actions).

Made ActionPlayer synthesize target actions in the case where you sent up an Action with a target--this should override any existing target you had.

Added a new RequestAction method to ClientInputSender that can be called from anywhere, and it hooked it up to Fernando's new Action/Emote UI Bars.