Feature/Tank Shield-Charge (#83)
Overview:
This implements the Tank character's first special ability, the "charged shield". While playing as a Tank character, press and hold the "1" key to charge up your shield. The shield stops charging if:

it fully charges on its own (after 1 second),
the player lets go of the "1" key
the character is hit by any attack
Once the action stops "charging up" (for any of the above reasons), it continues to provide damage-reduction for ActionDescription.EffectDurationSeconds seconds.

The amount of damage-reduction depends on how "charged up" the action got: it ranges from 50% to 100% damage reduction, although note that it is not a linear progression -- most of the extra protection comes near the end of the charge-up.

There is one other (pretty big) detail: if the Boss performs a Trample Attack and hits the tank character while this Action is at max-charge , the tank suffers no ill effects from the trample, and in fact causes the Boss to be stunned! This causes the boss to stand still, shaking his head, unable to act.

Implementation Details:

Adding this feature requred several new RPC calls: first, there are RPC calls to communicate when the player stops "charging up" the current Action. (This is likely the only scenario in this demo where the active Action will need further player input! It had to be wired all the way through to the ActionFX, so it can be displayed correctly.) Second, there is a new RPC call to abort the currently-playing ActionFX; this is needed when the boss is stunned, so that his trample action graphics end.
I integrated this "charging up" behavior into ClientCharacterVisualization by creating a new type of Input handler, a ChargingUpInput. I fleshed out the input-handler system a teeny bit, but it's still basically a stub. I just tried to make it not-worse. (Currently the only other use of the input-handler system is to aim the mage's AoE attack.)
Because the input-handler system is not yet wired into the new UI buttons, I didn't attempt to make the buttons work. This means you can't "charge up" the attack by clicking on the UI icon -- you must press the "1" key to do so. I can integrate charging-up into the UI buttons in this PR if desired ... I just didn't want to make this PR more complex than necessary.
To support damage reduction, I added an extremely simple "enchantment callback" system. This lets active Actions alter the value of gameplay calculations. I was able to reuse this system for the boss-trample-stunned check. (I also added an enchantment type for being healed, because it was the logical opposite of being damaged, but nothing currently uses it.)
Implementing the boss's stun was tricky: there was no way for the TrampleAction to replace itself with a StunnedAction. Trying to add a new Action to the queue during the trample's Update() would corrupt the action queue! To solve this, I added a new callback function that gets called after End(), which allows an Action to immediately replace itself with a new one. I would like feedback on this approach. It's the simplest solution I could think of, but it's a little weird.
Cancelling the trample also required new client code to cancel currently-playing ActionFX. The ActionFX interface already had a Cancel() function, but nothing could call it.
Replaced the TimedSelfDestruct component with new SpecialFXGraphic component; this is intended for use by particle-effects and other temporary graphics, and allows for more direct control over their lifespans. Marked the old component as "do not use". (Note that I also updated the boss's Trample graphics to use this new component.)
For testing, I once again set the Boss to use his Charge attack as his default attack. But note that I also fixed his "real" base-attack. The version that's in develop doesn't work: it had no melee expiration time, and its attack animation trigger was invalid. I fixed these problems, though, so when you want the boss to use his basic attack again, just change the enum on the ImpBoss character. Switch it to say "Imp Boss Base Attack" instead of "Imp Boss Trample Attack". You don't need need to change any Action data.
We can optionally have the boss suffer increased damage while stunned, although this isn't currently enabled in the data. Change the Amount value of the new Stunned action description to enable it. (I think setting it to 1.5, for 50% damage bonus, might be fun.)
I have set the shield's protective effect to last for 1.0 seconds, not 0.5 as mentioned in the jira task. (I found 0.5 to be frustratingly short, especially when trying to stun the boss.) This can be changed in the action description data.
Since there's no visual feedback on damage taken, I added a Debug.Log line to report it... but removed all debugging lines for this checkin. They might need to be re-added if testing