Feature: Archer charged shot (#102)
Adds the first special ability for the Archer class. Press and hold the "1" key to charge up the archer's power shot. (It stops charging up when you release the key, when you get hit, or when it's fully charged.) The projectile it fires will depend on how charged up it was: with minimal charge, the attack is pretty much just a basic attack, but with max-charge, the projectile does much higher damage, can fly farther, and can pierce through any number of enemies (that happen to be in the line of fire).

Edit: and if you get hit while charging up, you don't shoot the projectile.

This ability uses final artwork, so if you see artwork problems, they should be considered bugs, I guess!

Note that press-and-hold doesn't work for the GUI buttons yet, so you must press 1 on the keyboard to charge and release the attack.

Implementation Details
There isn't actually a ton of new code here; we already have the projectile logic and the charging-up-attacks logic, so most of this is refactoring (and the new art prefabs).

Note: added a behaviour Editor: I added a very simple Editor script to auto-fill out a list of ParticleSystems, because it was getting annoying to maintain the list by hand. Ironically I didn't end up needing this for the final artwork of this particular attack! But I left the code there since it will probably be needed for future attack graphics.

However, it uses the "old" Editor system (because it's less code), and I just stuck it in the same file as the actual class with #if UNITY_EDITOR blocks (because I find this method easier to maintain than keeping them in separate files). I'm concerned that some or all of these decisions will be considered inappropriate for this demo -- in which case, please let me know the appropriate way to do this! Or, in fact, I can just ditch it, since we don't need it at the moment. The Editor code is in SpecialFXGraphic.cs

Edit: oh also, I'm having trouble getting multiple clients to connect correctly right now. This isn't unique to this PR, it's a problem with the develop branch. (Even the long delays are no longer working.) So the networked aspects of this attack haven't been tested as well as I'd like. But a few simple tests worked, and there's nothing too new here, so it should be fine...

Other details:

Refactored ActionDescription to support a list of possible projectiles.
Refactored ServerProjectileLogic to support hitting multiple targets (i.e. piercing through a target to hit another one behind them)
Refactored ChargedShieldActionFX to work for both the shield-charge and the archer-charge, renamed class to ChargedActionFX
added new class derived from LaunchProjectileAction that supports "charging up" the projectile
Merging notes:

NetworkingManager prefab has three new Spawn()able network objects: the three tiers of charged-shot arrow
the only change made to MainMenu scene was to fix the in-scene NetworkingManager. (The in-scene instance was overriding the prefab's list of spawnable objects.)
Animation Controller was changed to add new states for archer's charge-up loop. No other changes