Feature/rogue stealth (#113)
This implements the rogue's Stealth Mode ability. Once the action's ExecTime has lapsed, the character becomes "stealthy". (If they are attacked during the ExecTime, the action aborts and they don't become stealthy.)

On the stealthy player's client, the character is enveloped in smoke. (The current smoke particles are junk programmer-art.) For other clients, stealthy characters are rendered as having no body parts (and hence are invisible). Monsters cannot perceive stealthy characters, and if they were pursuing a stealthy character, they give up.

Stealth ends when the rogue performs any type of attack, or gets hit by an attack (which can happen due to non-targeted damage, such as by being in the path of travel of the boss and getting accidentally trampled, or by being in the burst of an AoE).