More merge changes
Ran into a very weird problem with GameDataSource. The array entries for both the ArcherBaseAttack and MageBaseAttack were null. I tried plugging them back in, and I could plug one or the other in fine, but when I tried to plug BOTH into their respective slots, they went back to null! I couldn't figure out what was wrong and ultimately resolved it by giving one of them (ArcherBaseAttack) a new .meta file. Then I was able to plug both in. Since the file is only referenced by GameDataSource I think this shouldn't be a problem, but I wanted to document it in case somebody wondered.