Feature/parenting for animation-triggered FX
This is a simple enhancement to `AnimatorTriggeredSpecialFX` to let you specify what bone to parent a graphical prefab to (instead of being parented to the root GameObject).

This is needed for a few special effects, such as the "eyes glow red" effect of the boss, the Mage character's staff-glow effect, and various weapon-trail effects.

For completeness, I added the code to both the "on enter node" and "on exit node" logic, although we only intend to use it on the on-enter-node case.