Volume slider [GOMPS-373]
Adds a little slider for volume to the bottom-left of each UI screen. Stores this volume setting in PlayerPrefs.

Since we don't have an icon for volume handy in the release branch, it's just labeled with the word "Volume", but it's set up in a prefab that should let us change that to a speaker icon without having to alter the scenes again.

Because different UI screens have different scaling, it's hard to get the volume slider into the EXACT same spot on each screen without manually repositioning child elements -- and I wanted to avoid overriding any child-element positions, because because we want to replace the label with an icon later. So I changed the in-scene scale of the slider a little bit to make it approximately the same size and position on each screen. When final art is in, we can tweak position on each screen to get them to line up more precisely.