Feature/Rogue Dash Attack
Implements rogue's second special ability, the Dash Attack: the rogue dashes to the target spot and performs a melee attack. While dashing, the rogue cannot take damage.

This is implemented using the same "charge forward" logic as the Trample attack: the character moves in a straight line toward their target, as opposed to nav-mesh pathing. So if the player clicks on the other side of a wall, their character runs into the wall (and performs a harmless melee attack against the wall.)