Fix mage-bolt friendly fire [GOMPS-299]
Many actions can cause the player to "look like" they're hitting another player (e.g. selecting a target and using the Attack action on them makes them play the reaction anim) but that's just amusing fluff... the problem is that mage bolt did actual damage to other players. This is fixed.

Also fixed: if the target of the mage bolt had disappeared since the mage queued the attack, the attack was simply aborted because we had no target. Now, the attack is not aborted, and just turns into a normal "missed" bolt. To do this, the client sends the click-position along with all bolt attacks.