Feature/Show full particle on imp death
Imp monsters' GameObjects are being destroyed before their full "poof" particle can play. To address this, I added a new boolean that lets a particle be de-parented from the character.

Enabled this boolean for imps' death. Double-checked that the particle's `SpecialFXGraphics` has a timeout, so that it will self-destruct itself.

(At the moment, This is the only planned use of this new bool.)