Feature/remove cooldown time variable
Only one action was still using the CooldownTime feature: the boss's trample attack. Removed all vestiges of CooldownTime and changed the AI to instead check ReuseTime.

There was also a subtle side-effect to the boss's use of cooldown time: the boss would stand still for a few seconds after the trample ended. The boss no longer does this, and instead immediately tries to use new attacks. This required a change to the animation controller, because previously the boss would stay in his "trampling" pose until all movement stopped. Since he doesn't stop moving reliably anymore, we just use a stop-trigger to stop him. (Which we probably should have done in the first place, but it was early in the project and we were trying to minimize the number of triggers used for some reason)