Fixed DI issues
DI scope now reliably works with method and constructor based injection.
Renamed and refactored UpdateSlow into UpdateRunner - a tool for non-monobehs to run their update logic and for monobehs that want to save on an extra Update call.
Added Bootstrap.cs - a class where we set up core dependencies and that then proceeds to load the first logical scene.