fix: disconnect issue with wifi turned off (#562)
Disconnecting wifi caused an issue where clients wouldn't be able to reconnect.
This is due to a bug in NGO's Shutdown, which happens when it's called when NGO isn't Listening.

This PR adds a check around each Shutdown to make sure we're listing before calling it. --> workaround
This PR also cleans up our disconnect handling, to make sure we're using the same connection flow each time. --> cherrypick
This also fixes an issue where we didn't sent the appropriate allocation Id to lobby. --> cherrypick