feat: removing bridge classes for game states MTT-4204 (#697)
* Adjusted GameStateBehaviour class description

* Removed ClientBossRoomState

And renamed ServerBossRoomState to BossRoomState.
Added lifetime scopes to all the subscenes of BossRoom scene. These autoinject gameobjects that live under the "static network objects" groupings.

* Removed NetworkGameState and NetworkWinState

replaced with BufferedNetworkMessageChannel (an addition in this entry) - this message channel is identical to NetworkMessageChannel, but it inherits from a buffered message channel instead. This is used as a way to pass data that syncs from server to clients between different scenes without the need for SO variables, which are hard to understand.

* ClientMainMenuState renamed to MainMenuState

* Removed ClientPostGameState and renamed ServerPostGameState to PostGameState

also added auto-injected objects to the PostGameState-bearing game object.

* Changelog and typo fixes

* nullref fix (was disposing of a subscription handle before it was returned from the method)

* Removed BufferedNetworkMessageChannel

Replaced it's usage with WinState with a NetworkWinState that's preserved by BossRoomState to survive onto the next scene, PostGameState, which then destroys that preserved gameobject/NetworkWinState when we exit PostGameState to whatever next state.

* actually setting win state

* #

* Changed some type names based on feedback

* Cleanup

* Update Assets/Scripts/Gameplay/GameState/ServerBossRoomState.cs

Co-authored-by: Fernando Cortez <fernando.cortez@unity3d.com>

* Update Assets/Scripts/Gameplay/GameState/GameStateBehaviour.cs

Co-authored-by: Fernando Cortez <fernando.cortez@unity3d.com>

* Update CHANGELOG.md

Co-authored-by: Sam Bellomo <71790295+SamuelBellomo@users.noreply.github.com>

Co-authored-by: Fernando Cortez <fernando.cortez@unity3d.com>
Co-authored-by: Sam Bellomo <71790295+SamuelBellomo@users.noreply.github.com>