feat: action system cleanup MTT-4244 (#701)
* Adjusted GameStateBehaviour class description

* Removed ClientBossRoomState

And renamed ServerBossRoomState to BossRoomState.
Added lifetime scopes to all the subscenes of BossRoom scene. These autoinject gameobjects that live under the "static network objects" groupings.

* Removed NetworkGameState and NetworkWinState

replaced with BufferedNetworkMessageChannel (an addition in this entry) - this message channel is identical to NetworkMessageChannel, but it inherits from a buffered message channel instead. This is used as a way to pass data that syncs from server to clients between different scenes without the need for SO variables, which are hard to understand.

* ClientMainMenuState renamed to MainMenuState

* Removed ClientPostGameState and renamed ServerPostGameState to PostGameState

also added auto-injected objects to the PostGameState-bearing game object.

* moving files to different subfolders for better organization

* Changelog and typo fixes

* nullref fix (was disposing of a subscription handle before it was returned from the method)

* Removed BufferedNetworkMessageChannel

Replaced it's usage with WinState with a NetworkWinState that's preserved by BossRoomState to survive onto the next scene, PostGameState, which then destroys that preserved gameobject/NetworkWinState when we exit PostGameState to whatever next state.

* actually setting win state

* #

* further folder adjustments

* Namespace and funtion name changes for Actions

Moved all Action-related code into .Actions namespace, adjusted the rest of the codebase accordingly.
Renamed Start and Update methods of Actions to be OnStart and OnUpdate (to avoid confusing overlap with MonoBehavior event calls).
Moved FXProjectile to GameplayObjects folder.

* Start renamed to OnStart in ActionVisualization

* moving new actions appropriately

* Restored Client/Server names for game states

these were accidentally overriden during merges